enterprise-knowledge-copilot/requirements.txt
pdfplumber
faiss-cpu
python-dotenv
sentence-transformers
openai>=1.12.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
prometheus-client>=0.19.0
scikit-learn>=1.3.0


